+++ src/main/java/com/chat/main/application/chat/repository/ChatRoomDao.java
@@ -1,4 +1,6 @@
-package com.chat.main.application.chat.domain;
+package com.chat.main.application.chat.repository;
+
+import com.chat.main.application.chat.domain.ChatRoom;
 
 import java.util.List;
 import java.util.Optional;
@@ -8,4 +10,6 @@ public interface ChatRoomDao {
     void saveAll(List<ChatRoom> chatRoomList);
     List<ChatRoom> findByMemberId(Long memberId);
     Optional<ChatRoom> findByChatRoomByChatTypeAndTwoMembers(Long sendMemberId, Long targetMemberId);
+
+    Long createChatRoomId();
 }
